LayoutTests/inspector/debugger/sourceURLs-expected.txt

Tests for the Debugger.scriptParsed sourceURL and sourceMappingURL comment directive parsing.


== Running test suite: Debugger.scriptParsed.sourceURL-directives
-- Running test case: TestExpressionsForSourceURL
Found Script with sourceURL 'test1.js'
Found Script with sourceURL 'test2.js'
Found Script with sourceURL 'test3.js'
Found Script with sourceURL 'test4.js'
Found Script with sourceURL 'test5.js'
Found Script with sourceURL 'test6.js'
Found Script with sourceURL 'test7.js'
Found Script with sourceURL 'test8.js'
Found Script with sourceURL 'second.js'

-- Running test case: TestExpressionsForSourceMappingURL
Found Script with sourceMappingURL 'test1.js'
Found Script with sourceMappingURL 'test2.js'
Found Script with sourceMappingURL 'test3.js'
Found Script with sourceMappingURL 'test4.js'
Found Script with sourceMappingURL 'test5.js'
Found Script with sourceMappingURL 'test6.js'
Found Script with sourceMappingURL 'test7.js'
Found Script with sourceMappingURL 'test8.js'
Found Script with sourceMappingURL 'second.js'
